First prototype
Changed pin layout to better suit Veroboard construction.
Took out the SLEEP. It didn't work on the real device. I guess
due to global interrupts being off.

Tweaked thresholds as my battery sits at 12.9V even though it's been
on the shelf for a day or so. Upped float threshold to 13.0 and orange
threshold to somewhere between 70% and 100%. I'll have to look at the
discharge curve and work back.

MCLRE is enabled at the moment. I wondered if I was having problems with
slow start, something I've seen on other PICS. Seems not to be the case, so
I could turn that off when I want to use the pin for something.

My 5V regulators are of the cheaper 5% type. Next time I'll buy the more
expensive 0.5%. As a design point - maybe the extra cost of that accuracy
would be more than the extra cost of a better PIC. I happen to have a few
of these PICs on stock. My regulator is measuring 5.05V at the moment, so
1% error in line with the resistor chains I'm using.

I have 2mA LEDs with 2K2 resistors, so 1.5mA in practice. Not very bright,
but low drain on the battery.