Started work supporting bi-colour LED
In time for winter - our cycle lights have an old comparator based
monitor which drives a bi-colour LED. Replace this old circuit with
one based on the PIC Battery Monitor. If we can keep the software able
to drive both circuits then the choice is left to the constructor which
one to use.

The bi-colour LED is between GP0 and GP1 with GP0 high for green. A large
resistor pulls the pair down to ground so the microcontroller can determine
which circuit it is connected to.

The cycle lights use NiMH, so will need a different battery table. There
are a couple of possible fixes. My current plan is to allow more than one
battery table and allow the user to switch between them. GP2 remains
available for a push button as in the other circuit.